Refactor plugins to use addFiles plugin property
