--- packages/create-next-stack/src/main/plugins/mantine/setup/mantine-theme.ts
@@ -1,9 +1,0 @@
-import endent from "endent"
-
-export const mantineTheme = endent`
-  import { MantineThemeOverride } from "@mantine/core";
-
-  export const mantineTheme: MantineThemeOverride = {
-    colorScheme: "light",
-  };
-`
