--- packages/create-next-stack/src/main/plugins/sass/add-content/styles/global-styles.ts
@@ -1,27 +1,0 @@
-import endent from "endent"
-
-export const generateGlobalStyles = (): string => endent`
-  * {
-    box-sizing: border-box;
-  }
-
-  html,
-  body {
-    padding: 0;
-    margin: 0;
-    font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-      Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-    line-height: 1.5;
-  }
-
-  code {
-    font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono,
-      Bitstream Vera Sans Mono, Courier New, monospace;
-  }
-
-  a {
-    color: inherit;
-    text-decoration: none;
-    font-weight: bold;
-  }
-`
